Add the concept of HTTP session
The idea of a single session type doesn't quite work if BiDi wants to
reuse the WebDriver session, but also allow the possibility of having
(maybe multiple) BiDi-only sessions.

There are various ways we could address this e.g. we could have a
purportedly unrelated session type in WebDriver BiDi that happens to
share a session id when there's an associated HTTP-based session. But
the approach here is to assume a single session concept that's
composed of some universal state (e.g. the session id) and some
protocol-specific state.

To this end we define a HTTP flag on the session, and a HTTP Session
as one with the extra flag set. HTTP session have state associated
with the HTTP-based ("WebDriver classic") protocol which won't be used
in WebDriver-BiDi. The rest of the spec then operates on HTTP Sessions
exclusively. We also restrict the limitation of one session per
endpoint node to one HTTP session per endpoint node, allowing for the
possibility of more sessions that don't use the HTTP protocol.